# Deel_96-CookSmarter-1.0.0-swagger.yaml
openapi: 3.0.0
servers:
  # Added by API Auto Mocking Plugin
  # Added by API Auto Mocking Plugin
  - description: SwaggerHub API Auto Mocking
    url: https://virtserver.swaggerhub.com/Deel_96/CookSmarter/1.0.0
  - description: SwaggerHub API Auto Mocking
    url: https://cook-smarter-api.herokuapp.com/recipes2
info:
  description: CookSmarter API
  version: "1.0.0"
  title: API of the CookSmater project
  contact:
    email: dennis.eller@fh-bielefeld.de
  license:
    name: Apache 2.0
    url: 'http://www.apache.org/licenses/LICENSE-2.0.html'
...
paths:
...
  /login:
   post:
    tags:
      - authentification
    summary: Logging user in.
    operationId: loginUser
    responses:
      '201':
        description: Login sucessfull.
      '400':
        description: E-Mail or password wrong.
    requestBody:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Login'
      description: Login data.
...
  /recipes:
    get:
      tags:
        - recipes
      summary: Returns all recipes.
      operationId: getAllRecipes
      description: returns all recipes
      parameters:
        - in: query
          name: searchString
          description: optional searchstring for recipe name
          required: false
          schema:
            type: string
        - in: query
          name: skip
          description: number of records to skip for pagination
          schema:
            type: integer
            format: int32
            minimum: 0
        - in: query
          name: limit
          description: maximum number of records to return
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 50
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/RecipePreview'
        '400':
          description: bad input parameter
  /recipes/{recipeId}:
    get:
      tags:
        - recipes
      summary: Returns a single recipe.
      operationId: getRecipeById
      parameters:
        - in: path
          name: recipeId
          description: Id of a recipe
          required: true
          schema:
            type: integer
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Recipe'
        '400':
          description: Invalid ID supplied.
        '404':
          description: recipe not found.
...
  /me/recipes:
    get:
      tags:
        - recipes
      summary: Returns all recipes from the logged in user.
      operationId: getAllRecipesFromLoggedInUser
      parameters:
        - in: query
          name: searchString
          description: pass an optional search string for looking up recipes
          required: false
          schema:
            type: string
        - in: query
          name: skip
          description: number of records to skip for pagination
          schema:
            type: integer
            format: int32
            minimum: 0
        - in: query
          name: limit
          description: maximum number of records to return
          schema:
            type: integer
            format: int32
            minimum: 0
            maximum: 50
      responses:
        '200':
          description: search results matching criteria
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Recipe'
        '400':
          description: bad input parameter
    post:
      tags:
        - recipes
      summary: Adds a recipe.
      operationId: addRecipe
      description: Adds an recipe to the system.
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/Recipe'
        description: Recipe item to add
      responses:
        '201':
          description: item created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Recipe"
        '400':
          description: 'invalid input, object invalid'
...
components:
  schemas:
    Recipe:
          type: object
          required:
            - id
            - name
            - picture
            - directions
            - preparationtime
            - cookingtime
            - difficulty
            - dateposted
            - tags
            - author
            - rating
            - ingredients
          properties:
            id:
              type: integer
              format: int32
              example: 1
            name:
              type: string
              example: Butterbrot
            picture:
              type: string
              example: Bild
            directions:
              type: string
              example: Butter auf Brot schmieren
            preparationtime:
              type: integer
              example: 0
            cookingtime:
              type: integer
              example: 5
            difficulty:
              type: integer
              example: 1
            dateposted:
              type: string
              format: date
              example: "13.03.2021"
            author:
              type: string
              example: "Rudi79"
            tags:
              type: array
              items:
                $ref: '#/components/schemas/Tag'
            rating:
              type: integer
              format: float
              example: 4.3
            online:
              type: boolean
              example: true
            ingredients:
              type: array
              items:
                $ref: '#/components/schemas/RecipeEntry'
    RecipePreview:
          type: object
          required:
            - id
            - name
            - picture
            - tags
            - author
            - rating
          properties:
            id:
              type: integer
              format: int32
              example: 1
            name:
              type: string
              example: Butterbrot
            picture:
              type: string
              example: Butterbrot
            rating:
              type: integer
              format: float
              example: 4.3
            tags:
              type: array
              items:
                $ref: '#/components/schemas/Tag'
    Tag:
      type: object
      required:
        - id
        - name
        - description
      properties:
        id:
          type: integer
          format: int32
          example: 1
        name:
          type: string
          example: kalt
        description:
          type: string
          example: Speisen, die kalt serviert werden.
    Login:
      type: object
      required:
        - email
        - password
      properties:
        email:
          type: string
          format: email
          example: dennis@test.de
        password:
          type: string
          format: password
          example: secure123
...
    RecipeEntry:
      type: object
      required:
        - id
        - unit
        - amount
        - freshness
      properties:
        id:
          type: integer
          format: int32
          example: 1
        ingredient:
          type: string
          example: Brot
        unit:
          type: string
          example: g
        amount:
          type: integer
          example: 100
        freshness:
          type: integer
          format: float
          example: 4.3
...